+++ plutus-list/CHANGELOG.md
@@ -2,11 +2,16 @@
 
 This format is based on [Keep A Changelog](https://keepachangelog.com/en/1.0.0).
 
-## Unreleased
+## 2.1 -- 2022-02-06
+
+### Added
+
+* `PlutusTx.AssocMap.Natural` for `take`, `drop` and `splitAt` working on
+  `AssocMap`s
 
 ## 2.0 -- 2022-01-21
 
 ### Changed
-  
-* Plutus upgrade: `plutus` pinned to `65bad0fd53e432974c3c203b1b1999161b6c2dce`, 
+
+* Plutus upgrade: `plutus` pinned to `65bad0fd53e432974c3c203b1b1999161b6c2dce`,
   `plutus-apps` pinned to `34fe6eeff441166fee0cd0ceba68c1439f0e93d2`